+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Lesson metadata and cards.
 
+### Fixed
+- Unintended attack surfaces.
+
 ## [0.1.0] - 2021-11-10
 ### Added
 - Attack vector for clickjacking.