+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Unintended attack surfaces.
+- Unclosed file and response resources.
 
 ## [0.1.0] - 2021-11-10
 ### Added