+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.2.0] - 2026-09-12
 ### Added
 - Lesson metadata and cards.
 - Lesson card footer toggle.
@@ -42,5 +44,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Version reporting.
 - Docker base image.
 
-[Unreleased]: https://github.com/sgabe/DSVPWA/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/sgabe/DSVPWA/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/sgabe/DSVPWA/releases/tag/v0.2.0
 [0.1.0]: https://github.com/sgabe/DSVPWA/releases/tag/v0.1.0